code optimized

+++ ERPWebAppData/Entity/Loan.cs
@@ -11,13 +11,13 @@ namespace WebApp.Data.Entity
 
         public string LoanNumber { get; set; } = string.Empty;
 
-        public double LoanAmount { get; set; }
+        public decimal LoanAmount { get; set; }
 
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
 
         public int Tenure { get; set; }
 
-        public double EMI { get; set; }
+        public decimal EMI { get; set; }
 
         public bool IsReducingInterest { get; set; }
