+++ ERPWebAppModels/Transaction/LoanRequestModel.cs
@@ -7,11 +7,11 @@
 
         public string LoanNumber { get; set; } = string.Empty;
 
-        public double LoanAmount { get; set; }
+        public decimal LoanAmount { get; set; }
 
-        public double EMI { get; set; }
+        public decimal EMI { get; set; }
 
-        public double Rate { get; set; }
+        public decimal Rate { get; set; }
         public bool interestCalculationType { get; set; }
 
         [System.ComponentModel.DataAnnotations.Range(
